third_party/WebKit/LayoutTests/animations/underlying-background-position-edge.html

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
test(() => {
  target.style.backgroundPosition = 'right 20px bottom 20px';
  target.animate({backgroundPosition: 'left 20px top 20px'}, {iterationStart: 0.5, fill: 'forwards'});
  assert_equals(getComputedStyle(target).backgroundPosition, '50% 50%');
}, 'Don\'t ignore the specified edge of the underlying background-position value in animations');
</script>
